added facility to wrap entire process of compoent creation. changed LayerMemberFactory to also honor predefined id values. minor changes to id handling.

+++ file-driver-utils/src/main/java/de/ims/icarus2/filedriver/schema/resolve/ResolverContext.java
@@ -34,11 +34,11 @@ public interface ResolverContext {
 	Container currentContainer();
 
 	/**
-	 * Returns the index within the current container that reading is taking place
+	 * Returns the index within the current container that reading is taking place at
 	 *
 	 * @return
 	 */
-	int currentIndex();
+	long currentIndex();
 
 	/**
 	 * Returns the item with which currently read data should be associated
